auth

+++ backend/controllers/inventoryController.js
@@ -1,23 +1,24 @@
 const asyncHandler = require("express-async-handler")
 
 const Inventory = require("../models/inventoryModel")
+const User = require("../models/userModel")
 
 const {tuluStockNumberGenerator} = require("../middleware/stockNumberMiddleware")
 
 const {optionhighlights} = require("../middleware/vehicleOptionsMiddleware")
 
 // desc: gets inventory
-// route: GET /api/getInventory
+// route: GET /api/inventory
 // access: Private
 const getInventory = asyncHandler(async (req, res) => {
-    const car = await Inventory.find()
+    const car = await Inventory.find({user: req.user.id})
 
     res.status(200).json(car)
 })
 
 
 // desc: uploads inventory
-// route: POST /api/getInventory
+// route: POST /api/inventory
 // access: Private
 const addInventory = asyncHandler(async (req, res) => {
     if(!req.body.vin) {
@@ -29,6 +30,7 @@ const addInventory = asyncHandler(async (req, res) => {
     const highlightedOptions = optionhighlights(parsedOptions)
 
     const car = await Inventory.create({
+        user: req.user.id,
         dealerID: req.body.dealerID,
         tuluStockNum: stockNumber,
         dealerStockNumber: req.body.dealerStockNumber,
@@ -51,7 +53,7 @@ const addInventory = asyncHandler(async (req, res) => {
 })
 
 // desc: updates inventory
-// route: PUT /api/getInventory
+// route: PUT /api/inventory
 // access: Private
 const updateInventory = asyncHandler(async (req, res) => {
     const car = await Inventory.findById(req.params.id)
@@ -60,7 +62,21 @@ const updateInventory = asyncHandler(async (req, res) => {
         res.status(400)
         throw new Error("Car not found")
     }
+
+    const user = await User.findById(req.user.id)
+
+    // checking for user
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
     
+    // making sure logged in user matches the vehicle user
+    if(car.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
+    }
+
     const updatedInventory = await Inventory.findByIdAndUpdate(req.params.id, req.body)
     const updatedCar = await Inventory.findByIdAndUpdate(req.params.id, req.body) // running twice to get back the update document
 
@@ -68,7 +84,7 @@ const updateInventory = asyncHandler(async (req, res) => {
 })
 
 // desc: deletes inventory
-// route: DELETE /api/getInventory
+// route: DELETE /api/inventory
 // access: Private
 const deleteInventory = asyncHandler(async (req, res) => {
     const car = await Inventory.findById(req.params.id)
@@ -76,6 +92,20 @@ const deleteInventory = asyncHandler(async (req, res) => {
     if (!car) {
         res.status(400)
         throw new Error("Car not found")
+    }
+
+    const user = await User.findById(req.user.id)
+
+    // checking for user
+    if(!user) {
+        res.status(401)
+        throw new Error("User not found")
+    }
+    
+    // making sure logged in user matches the vehicle user
+    if(car.user.toString() !== user.id) {
+        res.status(401)
+        throw new Error("User not authorized")
     }
 
     await Inventory.findByIdAndRemove(req.params.id)
